databricks/demos/snowflake_iceberg_support/requirements.txt
databricks-connect==16.1.*
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
pyarrow<19.0.0
pyiceberg==0.9.0
setuptools>=60.0.0